[CI] Clean the statuses mess.
We have been leaving statueses as pending, the main reason is that when
we ran the tests, we set the wrong status. This was complicated because
we wrote comments AND set the statuses mixing two operations. We now
have the status and the comment separated for easy debugging and more
flexibility.