[CI] Add a status for the PR API diff.
To avoid issues when setting automerging we add a status for the API
diff. If human intervention is needed to check the changes the status
will be set to error, this way the automerging wont happen. We set the
status to error even when we only have a warning, better safe than
sorry.