[CI] Use template extension to simplify the separation between api diff and building the pkgs. (#14764)
Rather than having a single yaml template with several branches using ifs, we move to a single template that does the build and then this same template allows to perform a number of steps post build and before the clean up of the bot.

This allows to separate completely the steps used by the api diff and building the pkgs. The plan is to use this as the base to move the sining of the pkgs to a diff stage allowing the signing to run in parallel allowing to reduce the build time.