[CI] Move templates to their own dir. (#14853)
Now that we have parallel signing we can try and get each of the signing
and the notarizing in its own job. At the moment those steps are
sequential, we sign xamarin.ios, then xamarin.mac etc.. we can make an
improve by moving to diff jobs and make time to sign the build.zip and
msbuild.zip for XMA. This commit is to start the move towards that
approach.