[Nnyeah] Ensure that we have all the needed xamarin-macios dlls when using make. (#14910)
We need to be able to embed the dlls as a resource to compare the diff
assemblies. To do so we are going to ensure that we have called make in
the root dir so that the needed deps are present. This change is not yet
in the csproj, that will be a diff commit.



Co-authored-by: Chris Hamons <chris.hamons@xamarin.com>