[nnyeah] Refactor reworker to a 1-pass initialization
Reworker previous to this PR was very Lazy is it's init process. In
many ways this was beneficial, but was tricky to get right and made
nullability attributes a nightmare to add. It also included multiple "null objects"
that would show errors only at the end of a conversion when writing the file.

I replaced it with a single-pass init system, where the module is read
and all required attributes are fetched before processing begins.

Two requirements to consider:
- We don't want to ModuleDefinition.ReadModule to be repeated
- We don't want to fetch all of the attributes if reworking is not
necessary, but need the ModuleDefination to know that

A static factory method `CreateReworker` only returns a reworker
if work is required. This way, there is never a half-loaded state
to consider.

Other:
- Renamed some fields to match standard coding convention
- Also update integration makefile to include new required args