[nnyeah] Fix method hashing of generic matching non-generic due to Cecil limitation. (#14962)
This gets iOS integration tests passing

From what I can tell we are failing due to:

    public static new PHLivePhotoViewAppearance GetAppearance (UITraitCollection traits) {
    public static new PHLivePhotoViewAppearance GetAppearance<T> (UITraitCollection traits) where T: PHLivePhotoView {

having the same hash, as the generic constraints are not printed.

To solve this we take on a small (250ish) dependency on code from MonoMod.Common (MIT licensed). They have code which creates stable hash keys that include generic information.