[devops] Compute better base for comparing API diff. (#14963)
Previously we'd compare the tip of the PR branch with the commit before merge commit into the target branch created by GitHub.

This doesn't work in the following scenario:

    main  gh  pr
    
       |  G1 |          (merge commit created by GitHub)
       | / \ |
       |/   \|
       M2    |
       |     P2
       |    /|
       |   / |
       |  /  |
       | /   |
       |/    |
       M1    |
       |     |
       |     P1
       |    /
       |   /
       |  /
       | /
       |/

We'd end up comparing P2 with M2, which is not what we want, because M2 could
have API changes that would show up as missing in P2.

It's actually even worse than that, because we could be in the following
scenario:

    main  gh  pr
    
       |  G2 |          (merge commit created by GitHub when api diff started)
       | / \ |
       |/   \|
       M3    |
       |     P3
       |     |
       |  G1 |          (merge commit created by GitHub when build started)
       | / \ |
       |/   \|
       M2    |
       |     P2
       |    /|
       |   / |
       |  /  |
       | /   |
       |/    |
       M1    |
       |     |
       |     P1
       |    /
       |   /
       |  /
       | /
       |/

And we still want to P2 with M2, but we'd compare P2 with M3.

We want to compare P2 with M1 instead. This is done by asking git for the
merge-base between P2 and M2.