[nnyeah] Refactor module usage (#15048)
Clean up usage of modules in the app.
the var 'module' gets used all over the place. I'm going to refactor to fix up usage of imports and this will make things much easier to track.

In addition, there are a number of types that get used from module.TypeSystem.SomeType which I put into locals just to tidy up.