[msbuild] Fix computing the output path for SceneKitAsset items with custom Link metadata. Fixes #15104.
All files in a *.scnassets directory are included as SceneKitAsset items, but
when we process these files, Apple's tool take the *.scnassets directory as
input. This means that we use the first file inside a *.scnassets directory,
compute the path to the actual *.scnassets directory, and work with that path
from then on. The problem was that the Link metadata (if present) would be
copied from the file's metadata to the directory's metadata, and thus any
computations involving the Link metadata for the directory would be wrong.

Fix this by computing the correct Link metadata for the directory if the file
has it.

Also add a test.

Fixes https://github.com/xamarin/xamarin-macios/issues/15104.