[runtime] Work around broken dlopen in new OSes in Xcode 14.
dlopen is broken in the latest OS versions: if you pass the path to a symlink, dlopen
will (re)open the target file even if the target file is already loaded (presumably
using the real path and not the symlink).

I've worked around it by redirecting DllImports to libobjc.dylib and libSystem.dylib
to the corresponding real path, and also updating some of our constants to point
to the real path and not a symlink.

The workarounds only work for .NET, legacy Xamarin would need a fix in Mono to
not dlopen "libc.dylib" here (and instead load "libSystem.B.dylib"):
https://github.com/mono/mono/blob/f354099a6b61418a78717578f939deadb30df848/mono/metadata/w32file-unix.c#L4953

Ref: https://github.com/xamarin/maccore/issues/2585.