[CI] Remove the need of a variable to state if we are building a PR. (#15205)
Deciding if we build a PR or not used to be more complicated since we
had to make the diff between a CI build and a PR build. Now, since we
added diff pipelines we do not longer need to check any variable since
we can use a parameter.

This new fact makes the decision making simpler (although forces use to
add a new parameter in a few templates). The overall result is a simple
way to decide what can be used or not in the pipeline.

* Simplify logic.

* Add missing param.

* Fix the checkout for signing in the pr build.

* There is not need to sign in PR builds.

The signature is not needed for the tests and using -s in codesign means
that it is only valid in the machine that signed it.