[msbuild] Copy partial app manifests to build server if they exists on Windows. Fixes #15267.
The exclusion of partial app manifests happened here, but the commit doesn't explain why:

https://github.com/xamarin/xamarin-macios/commit/f4a4b232f8130a9172f6ac86c52a64c42d227bd5#diff-178de6110858688b9f7c2e8e57a873f5ac9498b355a456bfc18547ab2df876bc

It's certainly wrong if the partial app manifest is a part of the project, but
this exclusion might be because partial app manifests might be added as a part
of the build as well.

So change the logic to copy partial app manifests from Windows if they exist
there.

Fixes https://github.com/xamarin/xamarin-macios/issues/15267.