[dotnet] Support creating installers from archives
By doing this we can reuse the `_CreateInstaller` target from the VS for Mac Archive and Publishing feature.

The main change here is when publishing from an archive, we don't have the entitlements used to sign the app originally, so we need to retrieve them from the app bundle in order to include them when resigning the app.