[dotnet] There's no need to replicate Microsoft.NET.RuntimeIdentifierInference.targets anymore.
.NET changed their logic so that
Microsoft.NET.RuntimeIdentifierInference.targets is now included after our
default RuntimeIdentifier computation logic.

However, we need to still set SelfContained=true for app extensions, but limit
it to when there's a RuntimeIdentifier (i.e. not a universal app). For
Universal apps it will still be set for the build for each RuntimeIdentifier.