[workload] Update net6.0 KnownFrameworkReference
A `Xamarin.Shared.Sdk.MultiTarget.targets` file has been added to update
the ref/runtime pack versions associated with the .NET 6 SDK.  This file
will only be imported when using the .NET 6 SDK, and it should work
around the need to add new net6 and net7 versioned aliases of the Ref
and Runtime packs.  Adding this file to `AfterMicrosoftNETSdkTargets`
will ensure that it is imported last, after all KnownFrameworkReferences
that need updating are declared.