[release/6.0.4xx][msbuild] Take the .NET version into account when computing the illink.dll location.
We use illink.dll from the executing .NET version: if we're building with .NET
7, then we're using illink from .NET 7. This means we can't hard-code 'net6.0'
when computing the path to illink.dll for remote builds.

Also add a check to verify that the computed path exists, because otherwise
the error we get can be somewhat confusing.

Fixes https://devdiv.visualstudio.com/DefaultCollection/DevDiv/_workitems/edit/1611403.