Make package signing work with some platforms disabled.
Change package signing so that we only try to sign packages for platforms that aren't disabled.

This makes the 'Prepare packages' job green in Azure Devops.

The 'Prepare Release' job still fails in the 'Convert NuGet to MSI' step (for
the branch I've been working on), because the package names are too long for
MSIs. This should be resolved for a proper release branch (i.e. when this PR
is merged), because package names are much shorter in a release branch.

Example results: https://devdiv.visualstudio.com/DevDiv/_build/results?buildId=6711641&view=results

Backport of ##16028.