[actions] Get the head repository using the correct syntax. (#16064)
Get the head repository using the correct syntax when trying to publish
autoformatted changes. This action is a 'workflow_run' action, and not a
'pull_request' action, and as such we must look in the 'workflow_run' object
for the data we need.

This will hopefully make the action able to publish results back to a fork.