[autoformat] Extract some logic into its own script file.
We have some problems when autoformatting PRs and the PR branch isn't fully
up-to-date with regards to the target branch.

I believe this is what happens:

1. When a PR is created (or modified), GitHub Actions will merge the PR branch
   with the target branch, and parse/load the merged *.yml files.
2. Then when we run the autoformatter, we're working on the tip of the PR
   branch (and not the merged result).
3. This means that we were using the list of projects to autoformat from the
   merged branch, but exeuting on the PR branch. This resulted in spurious
   autoformatting, because the autoformatted would autoformat more code than
   expected.

The fix I'm implementing is to move the list of projects to autoformat to a
separate script in source code. That way we'll work upon the list of projects
as they show up in the PR branch, and not the merged results.