[tools] Always weak link the SensorKit framework. Fixes #9938.
The SensorKit framework isn't available on all devices (for instance iPads),
and as such we can't link with it strongly.

This seems to be a bug in Apple's toolchain, because Xcode runs into the same
problem if you try to use an app referencing SensorKit on an iPad.

Fixes https://github.com/xamarin/xamarin-macios/issues/9938.